Site header now uses Headroom.js to auto hide/show.
This also moves the content area top border back as it's currently
difficult to make this work on the sticky header while also keeping the
border-radius and having it all look good. This will likely be revisited
at a later date.

Note also that the displace component is now added to the Gin libraries
in the ambientimpact_base theme, so we're removing that from
omnipedia_site.info.yml.