Refactored compact header background to fix tearing during transitions:
Basically, having two blocks side by side both pretending to be one
whole block looks fine when they're static but it was just about
impossible to have them slide up and down at exactly the same rate, so
you could notice one a frame or two ahead of the other. This fixes that
by using one element for the whole background (the <header> element), so
even with multiple elements sliding at the same time, it's a lot less
obvious.